--- a/gapis/api/vulkan/doc.go
+++ b/gapis/api/vulkan/doc.go
@@ -12,5 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package vulkan implementes the API interface for the Vulkan graphics library.
+// Package vulkan implements the API interface for the Vulkan graphics library.
 package vulkan