--- a/gapis/api/vulkan/transform_wait_for_perfetto.go
+++ b/gapis/api/vulkan/transform_wait_for_perfetto.go
@@ -1,0 +1,148 @@
+// Copyright (C) 2020 Google Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package vulkan implementes the API interface for the Vulkan graphics library.
+
+package vulkan
+
+import (
+	"bytes"
+	"context"
+
+	"github.com/google/gapid/core/event/task"
+	"github.com/google/gapid/core/log"
+	"github.com/google/gapid/gapir"
+	"github.com/google/gapid/gapis/api"
+	"github.com/google/gapid/gapis/api/transform2"
+	"github.com/google/gapid/gapis/replay"
+	"github.com/google/gapid/gapis/replay/builder"
+	"github.com/google/gapid/gapis/service"
+	"github.com/google/gapid/gapis/trace"
+)
+
+var _ transform2.Transform = &waitForPerfetto{}
+
+type waitForPerfetto struct {
+	traceOptions  *service.TraceOptions
+	signalHandler *replay.SignalHandler
+	buffer        *bytes.Buffer
+}
+
+type FenceCallback = func(ctx context.Context, request *gapir.FenceReadyRequest)
+
+func newWaitForPerfetto(traceOptions *service.TraceOptions, signalHandler *replay.SignalHandler, buffer *bytes.Buffer) *waitForPerfetto {
+	return &waitForPerfetto{
+		traceOptions:  traceOptions,
+		signalHandler: signalHandler,
+		buffer:        buffer,
+	}
+}
+
+func (perfettoTransform *waitForPerfetto) RequiresAccurateState() bool {
+	return false
+}
+
+func (perfettoTransform *waitForPerfetto) BeginTransform(ctx context.Context, inputCommands []api.Cmd, inputState *api.GlobalState) ([]api.Cmd, error) {
+	return inputCommands, nil
+}
+
+func (perfettoTransform *waitForPerfetto) ClearTransformResources(ctx context.Context) {
+	// Do nothing
+}
+
+func (perfettoTransform *waitForPerfetto) EndTransform(ctx context.Context, inputCommands []api.Cmd, inputState *api.GlobalState) ([]api.Cmd, error) {
+	inputCommands = append(inputCommands, createVkDeviceWaitIdleCommandsForDevices(ctx, inputState)...)
+
+	// Melih TODO: We have to explain this magic id
+	waitForFenceCmd := createWaitForFence(ctx, api.CmdID(0x3ffffff), func(ctx context.Context, request *gapir.FenceReadyRequest) {
+		perfettoTransform.endOfTransformCallback(ctx, request)
+	})
+	inputCommands = append(inputCommands, waitForFenceCmd)
+	return inputCommands, nil
+}
+
+func (perfettoTransform *waitForPerfetto) TransformCommand(ctx context.Context, id api.CmdID, inputCommands []api.Cmd, inputState *api.GlobalState) ([]api.Cmd, error) {
+	if id != 0 {
+		return inputCommands, nil
+	}
+
+	inputCommands = append(inputCommands, createVkDeviceWaitIdleCommandsForDevices(ctx, inputState)...)
+	waitForFenceCmd := createWaitForFence(ctx, id, func(ctx context.Context, request *gapir.FenceReadyRequest) {
+		perfettoTransform.transformCallback(ctx, request)
+	})
+	inputCommands = append(inputCommands, waitForFenceCmd)
+	return inputCommands, nil
+}
+
+func (perfettoTransform *waitForPerfetto) transformCallback(ctx context.Context, request *gapir.FenceReadyRequest) {
+	errChannel := make(chan error)
+	signalHandler := perfettoTransform.signalHandler
+
+	go func() {
+		err := trace.TraceBuffered(ctx,
+			perfettoTransform.traceOptions.Device,
+			signalHandler.StartSignal,
+			signalHandler.StopSignal,
+			signalHandler.ReadyFunc,
+			perfettoTransform.traceOptions,
+			perfettoTransform.buffer)
+		if err != nil {
+			errChannel <- err
+		}
+		if !signalHandler.DoneSignal.Fired() {
+			signalHandler.DoneFunc(ctx)
+		}
+	}()
+
+	select {
+	case err := <-errChannel:
+		log.W(ctx, "Profiling error: %v", err)
+		return
+	case <-task.ShouldStop(ctx):
+		return
+	case <-signalHandler.ReadySignal:
+		return
+	}
+}
+
+func (perfettoTransform *waitForPerfetto) endOfTransformCallback(ctx context.Context, request *gapir.FenceReadyRequest) {
+	if !perfettoTransform.signalHandler.StopSignal.Fired() {
+		perfettoTransform.signalHandler.StopFunc(ctx)
+	}
+}
+
+func createVkDeviceWaitIdleCommandsForDevices(ctx context.Context, inputState *api.GlobalState) []api.Cmd {
+	cb := CommandBuilder{Thread: 0, Arena: inputState.Arena}
+	allDevices := GetState(inputState).Devices().All()
+
+	waitCmds := make([]api.Cmd, 0, len(allDevices))
+
+	// Wait for all queues in all devices to finish their jobs first.
+	for handle := range allDevices {
+		waitCmds = append(waitCmds, cb.VkDeviceWaitIdle(handle, VkResult_VK_SUCCESS))
+	}
+
+	return waitCmds
+}
+
+func createWaitForFence(ctx context.Context, id api.CmdID, callback FenceCallback) api.Cmd {
+	return replay.Custom{T: 0, F: func(ctx context.Context, s *api.GlobalState, b *builder.Builder) error {
+		fenceID := uint32(id)
+		b.Wait(fenceID)
+		tcb := func(p *gapir.FenceReadyRequest) {
+			callback(ctx, p)
+		}
+		return b.RegisterFenceReadyRequestCallback(fenceID, tcb)
+	}}
+}