It contains:

- Profile Replay and all the transforms that required for this replay type.

This will be followed by the other query types.

Currently, it contains code duplication as we need to keep old transform
system running but they will be removed when the transition is over.

For more information about the new transform system:
go/transform-refactor-first-stage

For Q&A:
go/agi-transform-review