@yalcinmelihyasin 2020-06-12
There are two non-obvious part of this that needs to be reviewed carefully. 

- profiling_layers.go -> transform_profiling_layers.go
- wait_for_perfetto.go -> transform_wait_for_perfetto.go

Please check if there are any unintended change. 

Also, I will run benchmark on this PR and report it to here.