@yalcinmelihyasin 2020-06-15
A more readable paste is in `5923962473152512`

Performance Characteristics for a bulky test trace:

### Master

Trace Time      Trace Size   Trace Frames   State Serialization   Trace Frame Time   Interactive
----------         ----------   ------------   -------------------   ----------------      -----------
15.030618035s   98900167     5              1.896128262s          12.287449ms        486.690432ms

Caching Done   Interaction    Max Memory   Before MEC Frame Time   Trailing Frame Time
------------      -----------    ----------   ---------------------   -----------------
9.236842584s   855.008918ms   0            22.185827ms             11.165738ms

### This PR:

Trace Time      Trace Size   Trace Frames   State Serialization   Trace Frame Time   Interactive
----------         ----------    ------------   -------------------    ----------------      -----------
15.031261818s   98828742     5              1.871542531s          11.393011ms        493.173701ms

Caching Done   Interaction    Max Memory   Before MEC Frame Time   Trailing Frame Time
------------       -----------    ----------          ---------------------        -----------------
9.755684269s   829.794254ms   0            22.707738ms             11.60843ms