@yalcinmelihyasin 2020-06-16
I have rerun the benchmark and updated the paste(5923962473152512)

```
//PR

Trace Time      Trace Size   Trace Frames   State Serialization   Trace Frame Time   Interactive
----------      ----------   ------------   -------------------   ----------------   -----------
48.044279079s   151587393    5              3.278612878s          328.418824ms       3.719196965s

Caching Done    Interaction    Max Memory   Before MEC Frame Time   Trailing Frame Time
------------    -----------    ----------   ---------------------   -----------------
27.319154996s   3.281800544s   1021124080   62.068303ms             139.382954ms

//////////

Trace Time      Trace Size   Trace Frames   State Serialization   Trace Frame Time   Interactive
----------      ----------   ------------   -------------------   ----------------   -----------
51.046844167s   156662479    5              3.440315864s          329.885689ms       4.490773381s

Caching Done    Interaction    Max Memory   Before MEC Frame Time   Trailing Frame Time
------------    -----------    ----------   ---------------------   -----------------
31.170464729s   3.498454666s   1208369688   65.188671ms             145.458619ms

//////////

Trace Time      Trace Size   Trace Frames   State Serialization   Trace Frame Time   Interactive
----------      ----------   ------------   -------------------   ----------------   -----------
51.047971149s   156957560    5              3.48888691s           362.900379ms       4.628750326s

Caching Done    Interaction    Max Memory   Before MEC Frame Time   Trailing Frame Time
------------    -----------    ----------   ---------------------   -----------------
30.873091097s   3.759797224s   1493793248   64.265395ms             147.481378ms

// Master

Trace Time      Trace Size   Trace Frames   State Serialization   Trace Frame Time   Interactive
----------      ----------   ------------   -------------------   ----------------   -----------
51.087617413s   157081315    5              3.358335804s          340.00094ms        4.492777062s

Caching Done    Interaction    Max Memory   Before MEC Frame Time   Trailing Frame Time
------------    -----------    ----------   ---------------------   -----------------
30.857340482s   3.536169681s   1246717208   65.636015ms             144.087233ms

//////////

Trace Time      Trace Size   Trace Frames   State Serialization   Trace Frame Time   Interactive
----------      ----------   ------------   -------------------   ----------------   -----------
48.058985732s   157332835    5              3.573214835s          327.530459ms       4.425905503s

Caching Done    Interaction    Max Memory   Before MEC Frame Time   Trailing Frame Time
------------    -----------    ----------   ---------------------   -----------------
28.621564671s   3.307703226s   1156268400   61.296922ms             147.805366ms

//////////

Trace Time      Trace Size   Trace Frames   State Serialization   Trace Frame Time   Interactive
----------      ----------   ------------   -------------------   ----------------   -----------
51.062939315s   156993834    5              3.504922745s          338.947929ms       4.679851215s

Caching Done    Interaction    Max Memory   Before MEC Frame Time   Trailing Frame Time
------------    -----------    ----------   ---------------------   -----------------
29.824837132s   3.546999745s   1388100776   64.991908ms             148.4789ms
```